backend/pyproject.toml
[tool.poetry]
name = "Test task 'Mortgage Calculator'"
version = "1.0.0"
description = ""
authors = ["Ilmir Valitov <valitovilmir29@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
psycopg2-binary = "^2.9.3"

# Django
Django = "==3.1.6"

# Django Rest Framework
djangorestframework = "3.12.0"
drf-spectacular = '0.13.2'
coreapi = "2.3.3 "

# third
requests = "~2.23.0"
cryptography = "2.1.4"
django-filter = "21.1"
django-debug-toolbar = "2.2"

# Testing
factory-boy = "~2.12.0"
coverage = "~5.1.0"

# Setting codestyle
[tool.black]
line-length = 100

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

